fix test_struct_with_a_slice

+++ pkg/dynamicinstrumentation/testutil/fixtures.go
@@ -447,7 +447,7 @@ var structCaptures = fixtures{
 								*capturedValue("uint8", "4"),
 							},
 						},
-						"z": capturedValue("int", "5"),
+						"z": capturedValue("uint64", "5"),
 					},
 				},
 			},
